+++ training/train_crystal.py
@@ -22,13 +22,13 @@ from kgcnn.utils.devices import set_devices_gpu
 
 # Input arguments from command line.
 parser = argparse.ArgumentParser(description='Train a GNN on a CrystalDataset.')
-parser.add_argument("--model", required=False, help="Graph model to train.", default="MEGAN")
+parser.add_argument("--model", required=False, help="Graph model to train.", default="Schnet")
 parser.add_argument("--dataset", required=False, help="Name of the dataset or leave empty for custom dataset.",
                     default="MatProjectPhononsDataset")
 parser.add_argument("--hyper", required=False, help="Filepath to hyper-parameter config file (.py or .json).",
                     default="hyper/hyper_mp_phonons.py")
 parser.add_argument("--make", required=False, help="Name of the make function or class for model.",
-                    default="make_model")
+                    default="make_crystal_model")
 parser.add_argument("--gpu", required=False, help="GPU index used for training.",
                     default=None, nargs="+", type=int)
 parser.add_argument("--fold", required=False, help="Split or fold indices to run.",
@@ -130,23 +130,21 @@ for i, (train_index, test_index) in enumerate(kf.split(X=np.zeros((data_length, 
     if "scaler" in hyper["training"]:
         print("Using StandardScaler.")
         if hyper["training"]["scaler"]["class_name"] == "QMGraphLabelScaler":
-            scaler = QMGraphLabelScaler(**hyper["training"]["scaler"]["config"]).fit(y_train, atoms_train)
-            y_train = scaler.fit_transform(y_train, atoms_train)
-            y_test = scaler.transform(y_test, atoms_test)
-            scaler_shape = np.expand_dims(scaler.scale_, axis=0).shape
+            scaler = QMGraphLabelScaler(**hyper["training"]["scaler"]["config"])
         else:
             scaler = StandardScaler(**hyper["training"]["scaler"]["config"])
-            y_train = scaler.fit_transform(y_train)
-            y_test = scaler.transform(y_test)
-            scaler_shape = (1, 1)
+
+        y_train = scaler.fit_transform(X=y_train, atomic_number=atoms_train)
+        y_test = scaler.transform(X=y_test, atomic_number=atoms_test)
+        scaler_scale = scaler.get_scaling()
 
         # If scaler was used we add rescaled standard metrics to compile, since otherwise the keras history will not
         # directly log the original target values, but the scaled ones.
-        mae_metric = ScaledMeanAbsoluteError(scaler_shape, name="scaled_mean_absolute_error")
-        rms_metric = ScaledRootMeanSquaredError(scaler_shape, name="scaled_root_mean_squared_error")
+        mae_metric = ScaledMeanAbsoluteError(scaler_scale.shape, name="scaled_mean_absolute_error")
+        rms_metric = ScaledRootMeanSquaredError(scaler_scale.shape, name="scaled_root_mean_squared_error")
         if scaler.scale_ is not None:
-            mae_metric.set_scale(np.expand_dims(scaler.scale_, axis=0))
-            rms_metric.set_scale(np.expand_dims(scaler.scale_, axis=0))
+            mae_metric.set_scale(scaler_scale)
+            rms_metric.set_scale(scaler_scale)
         metrics = [mae_metric, rms_metric]
     else:
         print("Not using StandardScaler.")