+++ changelog.md
@@ -18,6 +18,9 @@ SortEdgeIndices(in_place=True)(GraphDict())
 * Refactored ``kgcnn.data.qm`` including all qm7-qm9 datasets. Improved documentation. If error occurs, please run `QM9Dataset(reload=True)`.
 * Refactored ``kgcnn.data.moleculenet``. Interface and behaviour does not change.
 * Renamed ``kgcnn.mol.graph_babel`` and ``kgcnn.mol.graph_rdkit`` and move conversion into ``kgcnn.mol.convert``.
+* Added ``kgcnn.data.datasets.MD17Dataset`` and ``kgcnn.data.datasets.MD17RevisedDataset``
+* Refactored ``kgcnn.scaler`` module to follow sklearn definitions. Changed input naming and order for scaler. Add config and weights functionality.
+
 
 v2.1.0